change from old main_oop file to new oop implementation
